requirements.txt
beautifulsoup4==4.9.0
boto3==1.12.39
botocore==1.15.39
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
distro==1.5.0
docutils==0.15.2
geographiclib==1.50
geopy==1.21.0
idna==2.9
jmespath==0.9.5
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
tabula-py==2.1.0
urllib3==1.25.8
